requirements.txt
alembic==1.18.0
bcrypt==4.0.1
email-validator==2.3.0
fastapi==0.119.0
fastapi-cache2==0.2.2
httpx==0.28.1
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.12.2
pydantic-settings==2.12.0
pytest==8.4.2
python-jose==3.5.0
python-multipart==0.0.20
redis==4.6.0
reportlab==4.4.10
slowapi==0.1.9
SQLAlchemy==2.0.43
starlette==0.48.0
uvicorn==0.35.0
